editing interface is working again
