+++ packages/relative-values/src/values/cache/build.ts
@@ -4,73 +4,15 @@
 import fs from "fs";
 import path from "path";
 
-import { getModelCode, Model, modelFromJSON } from "@/model/utils";
+import { Model, modelFromJSON } from "@/model/utils";
 import { Catalog, InterfaceWithModels } from "@/types";
 import { allInterfaces } from "@models/src/index";
-import { sq, SqLambda, SqProject } from "@quri/squiggle-lang";
-import { RelativeValueResult } from "../types";
+import { sq } from "@quri/squiggle-lang";
 import { CatalogCache, ModelCache } from "../cache";
+import { ModelEvaluator } from "../ModelEvaluator";
+import { RelativeValueResult } from "../types";
 
 const cacheDir = path.join(path.dirname(__filename), "../../../models/cache");
-
-const wrapper = sq`
-{|x, y|
-  dist = fn(x, y) -> SampleSet.fromDist
-  {
-    median: inv(dist, 0.5),
-    min: inv(dist, 0.05),
-    max: inv(dist, 0.95),
-    db: 10 * (SampleSet.map(dist, {|x| max([abs(x), 1e-9])}) -> log10 -> stdev)
-  }
-}
-`;
-
-function buildRelativeValue({
-  fn,
-  id1,
-  id2,
-}: {
-  fn: SqLambda;
-  id1: string;
-  id2: string;
-}): RelativeValueResult {
-  const result = fn.call([id1, id2]);
-  if (!result.ok) {
-    return { ok: false, value: result.value.toString() };
-  }
-  const record = result.value.asJS();
-  if (!(record instanceof Map)) {
-    return { ok: false, value: "Expected record" };
-  }
-
-  const median = record.get("median");
-  const min = record.get("min");
-  const max = record.get("max");
-  const db = record.get("db");
-
-  if (typeof median !== "number") {
-    return { ok: false, value: "Expected median to be a number" };
-  }
-  if (typeof min !== "number") {
-    return { ok: false, value: "Expected min to be a number" };
-  }
-  if (typeof max !== "number") {
-    return { ok: false, value: "Expected max to be a number" };
-  }
-  if (typeof db !== "number") {
-    return { ok: false, value: "Expected db to be a number" };
-  }
-
-  return {
-    ok: true,
-    value: {
-      median,
-      min,
-      max,
-      db,
-    },
-  };
-}
 
 function sleep(ms: number) {
   return new Promise((resolve) => setTimeout(resolve, ms));
@@ -83,35 +25,31 @@ async function modelToJson({
   catalog: Catalog;
   model: Model;
 }): Promise<ModelCache["relativeValues"]> {
-  const project = SqProject.create();
-  project.setSource("wrapper", wrapper);
-  project.setContinues("wrapper", ["model"]);
-  project.setSource("model", getModelCode(model));
-
-  project.run("wrapper");
-  const result = project.getResult("wrapper");
-
   let obj: Record<string, Record<string, RelativeValueResult>> = {};
-  if (result.ok && result.value.tag === "Lambda") {
-    for (const start of catalog.items) {
-      let row: Record<string, RelativeValueResult> = {};
 
-      for (const end of catalog.items) {
-        const relativeValue = buildRelativeValue({
-          fn: result.value.value,
-          id1: start.id,
-          id2: end.id,
-        });
-        if (!relativeValue.ok) {
-          console.error(
-            `Relative value returned a failure in model ${model.title}: ${relativeValue.value}. Comparing ${start.id} vs. ${end.id}`
-          );
-        }
-        await sleep(0); //Used to make sure that it's possible to interrupt if wanted by console user.
-        row[end.id] = relativeValue;
+  const modelResult = ModelEvaluator.create(model);
+  if (!modelResult.ok) {
+    console.error(`Failed to initialize a model: ${modelResult.value}`);
+    return obj;
+  }
+
+  for (const start of catalog.items) {
+    let row: Record<string, RelativeValueResult> = {};
+
+    for (const end of catalog.items) {
+      const relativeValue = modelResult.value.compareWithoutCache(
+        start.id,
+        end.id
+      );
+      if (!relativeValue.ok) {
+        console.error(
+          `Relative value returned a failure in model ${model.title}: ${relativeValue.value}. Comparing ${start.id} vs. ${end.id}`
+        );
       }
-      obj[start.id] = row;
+      await sleep(0); //Used to make sure that it's possible to interrupt if wanted by console user.
+      row[end.id] = relativeValue;
     }
+    obj[start.id] = row;
   }
   return obj;
 }
